Made admin page

+++ DiscussIt_Website/routes/web.php
@@ -16,11 +16,12 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', function () {
     return view('welcome');
 });
-Route::get('/posts', '\App\Http\Controllers\PostController@index');
-Route::get('/singlepost/{id}', '\App\Http\Controllers\PostController@show');
-Route::get('/create', '\App\Http\Controllers\PostController@create');
-Route::post('/posts', '\App\Http\Controllers\PostController@store');
-Route::delete('/posts/{id}', '\App\Http\Controllers\PostController@destroy');
+Route::get('/posts', '\App\Http\Controllers\PostController@index')->middleware('auth');
+Route::get('/admin', function (){return view('admin');});
+Route::get('/singlepost/{id}', '\App\Http\Controllers\PostController@show')->middleware('auth');
+Route::get('/create', '\App\Http\Controllers\PostController@create')->middleware('auth');
+Route::post('/posts', '\App\Http\Controllers\PostController@store')->middleware('auth');
+Route::delete('/posts/{id}', '\App\Http\Controllers\PostController@destroy')->middleware('auth');
 
 
 Auth::routes();
